water function
can enter water when umbrella is up. sprite will change. when umbrella
is down you are respawned to outside the water. right now the position
of being respawned is hard coded since theres only one body of water.
will try to work on respawning to the left / right of body of water
depending on relative location but the hardcoded value should suffice
for tomorrow!