Poking event switch
Must be in "poking" mode to interact with event switch. Right now that
code is to simply destroy the event switch when they interact, but when
we have an idea of what this event should be, that code can be to carry
out the event listed in the event switch script instead of "destroy the
event switch".
It looks like the best way to implement separate events as of now is to
create a short script for each event switch, where the script has a
method that will be executed when the player pokes it.