Trigger Overhaul
Except for nimboids, the way that all triggers work has been changed.
They are now designated by tag rather than name. This makes it much
easier to, for example, pick out which updraft you want to use in a
script, because you can just rename it. Triggers now have their own
script attached to them, which the playerscript calls, rather than
everything just being in the playerscript. Hide is now on S and pressing
W while the umbrella is open uses the new sprite. However, the
functionality of that is not yet implemented, and currently the game
still treats it like hiding.